Crate re_view_dataframe

Rerun Data View

A View that shows the data contained in entities in a table.

Modules

dataframe_ui 🔒

expanded_rows 🔒

view_class 🔒

view_query 🔒

visualizer_system 🔒

Structs

DataframeView

Query

Wrapper over the DataframeQuery blueprint archetype that can also display some UI.