re_view_dataframe::dataframe_ui
Function split_ui_vertically

fn split_ui_vertically<Item, Ctx>(
    ui: &mut Ui,
    context: &mut Ctx,
    line_data: impl Iterator<Item = Item>,
    line_content_ui: impl Fn(&mut Ui, &mut Ctx, usize, Item),
)

Helper to draw individual lines into an expanded cell in a table.

context: whatever mutable context is necessary for the line_content_ui line_data: the data to be displayed in each line line_content_ui: the function to draw the content of each line