re_view_dataframe::view_query::ui
Function all_pov_entities_for_view

fn all_pov_entities_for_view(
    ctx: &ViewerContext<'_>,
    view_id: ViewId,
    timeline: &TimelineName,
) -> BTreeSet<EntityPath>

Gather all entities that can meaningfully be used as point-of-view for this view.

Meaning:

the entity is part of this view

the entity has any component on the chosen timeline